backend/requirements.txt
Django==6.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.5
Pillow==10.1.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
PyMySQL==1.1.0
django-allauth

